split boolean logic library into files
